Update tests to use toContain instead of toMatchInlineSnapshot
- Updated hashbang-still-works.test.ts to use expect().toContain("error: Syntax Error")
- Removed duplicate single-hash-bounds-check.test.ts file
- All bounds check tests are now consolidated in hashbang-still-works.test.ts
